[REFACTOR] Novel 내에 존재하는 서재 로직을 분리
