+++ src/main/java/org/websoso/WSSServer/novel/service/NovelService.java
@@ -1,90 +1,39 @@
 package org.websoso.WSSServer.novel.service;
 
-import static org.websoso.WSSServer.domain.common.ReadStatus.QUIT;
-import static org.websoso.WSSServer.domain.common.ReadStatus.WATCHED;
-import static org.websoso.WSSServer.domain.common.ReadStatus.WATCHING;
-import static org.websoso.WSSServer.exception.error.CustomGenreError.GENRE_NOT_FOUND;
-import static org.websoso.WSSServer.exception.error.CustomNovelError.NOVEL_NOT_FOUND;
-import static org.websoso.WSSServer.exception.error.CustomUserNovelError.ALREADY_INTERESTED;
-import static org.websoso.WSSServer.exception.error.CustomUserNovelError.NOT_INTERESTED;
 
+import static org.websoso.WSSServer.exception.error.CustomNovelError.NOVEL_NOT_FOUND;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Random;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.websoso.WSSServer.domain.Avatar;
 import org.websoso.WSSServer.feed.domain.Feed;
-import org.websoso.WSSServer.domain.Genre;
-import org.websoso.WSSServer.domain.GenrePreference;
-import org.websoso.WSSServer.library.domain.Keyword;
 import org.websoso.WSSServer.novel.domain.Novel;
-import org.websoso.WSSServer.novel.domain.NovelGenre;
 import org.websoso.WSSServer.domain.PopularNovel;
-import org.websoso.WSSServer.domain.User;
-import org.websoso.WSSServer.library.domain.UserNovel;
-import org.websoso.WSSServer.library.domain.UserNovelKeyword;
-import org.websoso.WSSServer.domain.common.AttractivePointName;
-import org.websoso.WSSServer.domain.common.GenreName;
-import org.websoso.WSSServer.dto.keyword.KeywordCountGetResponse;
-import org.websoso.WSSServer.dto.novel.FilteredNovelsGetResponse;
-import org.websoso.WSSServer.dto.novel.NovelGetResponseBasic;
-import org.websoso.WSSServer.dto.novel.NovelGetResponseInfoTab;
-import org.websoso.WSSServer.dto.novel.NovelGetResponsePreview;
-import org.websoso.WSSServer.dto.novel.SearchedNovelsGetResponse;
-import org.websoso.WSSServer.dto.platform.PlatformGetResponse;
 import org.websoso.WSSServer.dto.popularNovel.PopularNovelGetResponse;
 import org.websoso.WSSServer.dto.popularNovel.PopularNovelsGetResponse;
-import org.websoso.WSSServer.dto.userNovel.TasteNovelGetResponse;
-import org.websoso.WSSServer.dto.userNovel.TasteNovelsGetResponse;
-import org.websoso.WSSServer.exception.exception.CustomGenreException;
 import org.websoso.WSSServer.exception.exception.CustomNovelException;
-import org.websoso.WSSServer.exception.exception.CustomUserNovelException;
 import org.websoso.WSSServer.repository.AvatarRepository;
 import org.websoso.WSSServer.feed.repository.FeedRepository;
 import org.websoso.WSSServer.repository.GenrePreferenceRepository;
-import org.websoso.WSSServer.novel.repository.NovelGenreRepository;
-import org.websoso.WSSServer.novel.repository.NovelPlatformRepository;
 import org.websoso.WSSServer.novel.repository.NovelRepository;
 import org.websoso.WSSServer.repository.PopularNovelRepository;
-import org.websoso.WSSServer.library.repository.UserNovelAttractivePointRepository;
-import org.websoso.WSSServer.library.repository.UserNovelKeywordRepository;
 import org.websoso.WSSServer.library.repository.UserNovelRepository;
-import org.websoso.WSSServer.service.GenreService;
-import org.websoso.WSSServer.library.service.KeywordService;
-import org.websoso.WSSServer.library.service.UserNovelService;
 
 @Service
 @RequiredArgsConstructor
 @Transactional
 public class NovelService {
 
-    private static final int ATTRACTIVE_POINT_SIZE = 3;
-    private static final int KEYWORD_SIZE = 5;
-
     private final NovelRepository novelRepository;
-    private final UserNovelService userNovelService;
-    private final UserNovelRepository userNovelRepository;
-    private final NovelPlatformRepository novelPlatformRepository;
-    private final UserNovelAttractivePointRepository userNovelAttractivePointRepository;
     private final FeedRepository feedRepository;
-    private final NovelGenreRepository novelGenreRepository;
-    private final UserNovelKeywordRepository userNovelKeywordRepository;
-    private final KeywordService keywordService;
-    private final GenreService genreService;
     private final PopularNovelRepository popularNovelRepository;
     private final AvatarRepository avatarRepository;
-    private final GenrePreferenceRepository genrePreferenceRepository;
 
     @Transactional(readOnly = true)
     public Novel getNovelOrException(Long novelId) {
@@ -93,271 +42,6 @@ public class NovelService {
                         "novel with the given id is not found"));
     }
 
-    //이건 상위 계층에서 처리
-    @Transactional(readOnly = true)
-    public NovelGetResponseBasic getNovelInfoBasic(User user, Long novelId) {
-        Novel novel = getNovelOrException(novelId);
-        List<NovelGenre> novelGenres = novelGenreRepository.findAllByNovel(novel);
-        Integer novelRatingCount = userNovelRepository.countByNovelAndUserNovelRatingNot(novel, 0.0f);
-        Float novelRating = novelRatingCount == 0 ? 0.0f
-                : Math.round(userNovelRepository.sumUserNovelRatingByNovel(novel) / novelRatingCount * 10.0f) / 10.0f;
-        return NovelGetResponseBasic.of(
-                novel,
-                userNovelService.getUserNovelOrNull(user, novel),
-                getNovelGenreNames(novelGenres),
-                getRandomNovelGenreImage(novelGenres),
-                userNovelRepository.countByNovelAndIsInterestTrue(novel),
-                novelRating,
-                novelRatingCount,
-                feedRepository.countByNovelId(novelId)
-        );
-    }
-
-    private String getNovelGenreNames(List<NovelGenre> novelGenres) {
-        return novelGenres.stream()
-                .map(novelGenre -> getKoreanGenreName(novelGenre.getGenre().getGenreName()))
-                .collect(Collectors.joining("/"));
-    }
-
-    private String getKoreanGenreName(String englishGenreName) {
-        return Arrays.stream(GenreName.values())
-                .filter(genreName -> genreName.getLabel().equalsIgnoreCase(englishGenreName))
-                .findFirst()
-                .map(GenreName::getKoreanLabel)
-                .orElseThrow(
-                        () -> new CustomGenreException(GENRE_NOT_FOUND, "Genre with the given genreName is not found"));
-    }
-
-    private String getRandomNovelGenreImage(List<NovelGenre> novelGenres) {
-        Random random = new Random();
-        return novelGenres.get(random.nextInt(novelGenres.size())).getGenre().getGenreImage();
-    }
-
-    public void registerAsInterest(User user, Long novelId) {
-        Novel novel = getNovelOrException(novelId);
-        UserNovel userNovel = userNovelService.getUserNovelOrNull(user, novel);
-
-        if (userNovel != null && userNovel.getIsInterest()) {
-            throw new CustomUserNovelException(ALREADY_INTERESTED, "already registered as interested");
-        }
-
-        if (userNovel == null) {
-            try {
-                userNovel = userNovelService.createUserNovelByInterest(user, novel);
-            } catch (DataIntegrityViolationException e) {
-                userNovel = userNovelService.getUserNovelOrException(user, novelId);
-            }
-        }
-
-        userNovel.setIsInterest(true);
-    }
-
-    public void unregisterAsInterest(User user, Long novelId) {
-        UserNovel userNovel = userNovelService.getUserNovelOrException(user, novelId);
-
-        if (!userNovel.getIsInterest()) {
-            throw new CustomUserNovelException(NOT_INTERESTED, "not registered as interest");
-        }
-
-        userNovel.setIsInterest(false);
-
-        if (isUserNovelOnlyByInterest(userNovel)) {
-            userNovelRepository.delete(userNovel);
-        }
-
-    }
-
-    private Boolean isUserNovelOnlyByInterest(UserNovel userNovel) {
-        return userNovel.getStatus() == null;
-    }
-
-    @Transactional(readOnly = true)
-    public NovelGetResponseInfoTab getNovelInfoInfoTab(Long novelId) {
-        Novel novel = getNovelOrException(novelId);
-        return NovelGetResponseInfoTab.of(
-                novel,
-                getPlatforms(novel),
-                getAttractivePoints(novel),
-                getKeywords(novel),
-                userNovelRepository.countByNovelAndStatus(novel, WATCHING),
-                userNovelRepository.countByNovelAndStatus(novel, WATCHED),
-                userNovelRepository.countByNovelAndStatus(novel, QUIT)
-        );
-    }
-
-    private List<PlatformGetResponse> getPlatforms(Novel novel) {
-        return novelPlatformRepository.findAllByNovel(novel).stream()
-                .map(PlatformGetResponse::of)
-                .collect(Collectors.toList());
-    }
-
-    private List<String> getAttractivePoints(Novel novel) {
-        Map<String, Integer> attractivePointMap = makeAttractivePointMapExcludingZero(novel);
-
-        if (attractivePointMap.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        return getTOP3AttractivePoints(attractivePointMap);
-    }
-
-    private Map<String, Integer> makeAttractivePointMapExcludingZero(Novel novel) {
-        Map<String, Integer> attractivePointMap = new HashMap<>();
-
-        for (AttractivePointName point : AttractivePointName.values()) {
-            attractivePointMap.put(point.getLabel(),
-                    userNovelAttractivePointRepository.countByUserNovel_NovelAndAttractivePoint_AttractivePointName(
-                            novel, point.getLabel()));
-        }
-
-        attractivePointMap.entrySet().removeIf(entry -> entry.getValue() == 0);
-
-        return attractivePointMap;
-    }
-
-    private List<String> getTOP3AttractivePoints(Map<String, Integer> attractivePointMap) {
-        Map<Integer, List<String>> groupedByValue = groupAttractivePointByValue(attractivePointMap);
-
-        List<String> result = new ArrayList<>();
-        List<Integer> sortedKeys = new ArrayList<>(groupedByValue.keySet());
-        Collections.sort(sortedKeys, Collections.reverseOrder());
-
-        Random random = new Random();
-
-        for (Integer key : sortedKeys) {
-            List<String> items = groupedByValue.get(key);
-            if (result.size() + items.size() > ATTRACTIVE_POINT_SIZE) {
-                Collections.shuffle(items, random);
-                items = items.subList(0, ATTRACTIVE_POINT_SIZE - result.size());
-            }
-            result.addAll(items);
-            if (result.size() >= ATTRACTIVE_POINT_SIZE) {
-                break;
-            }
-        }
-
-        return result;
-    }
-
-    private Map<Integer, List<String>> groupAttractivePointByValue(Map<String, Integer> attractivePointMap) {
-        Map<Integer, List<String>> groupedByValue = new HashMap<>();
-
-        for (Map.Entry<String, Integer> entry : attractivePointMap.entrySet()) {
-            groupedByValue
-                    .computeIfAbsent(entry.getValue(), k -> new ArrayList<>())
-                    .add(entry.getKey());
-        }
-
-        return groupedByValue;
-    }
-
-    private List<KeywordCountGetResponse> getKeywords(Novel novel) {
-        List<UserNovelKeyword> userNovelKeywords = userNovelKeywordRepository.findAllByUserNovel_Novel(novel);
-
-        if (userNovelKeywords.isEmpty()) {
-            return Collections.emptyList();
-        }
-
-        Map<Keyword, Long> keywordFrequencyMap = userNovelKeywords.stream()
-                .collect(Collectors.groupingBy(UserNovelKeyword::getKeyword, Collectors.counting()));
-
-        return keywordFrequencyMap.entrySet().stream()
-                .sorted(Map.Entry.<Keyword, Long>comparingByValue().reversed())
-                .limit(KEYWORD_SIZE)
-                .map(entry -> KeywordCountGetResponse.of(entry.getKey(), entry.getValue().intValue()))
-                .collect(Collectors.toList());
-    }
-
-    @Transactional(readOnly = true)
-    public FilteredNovelsGetResponse getFilteredNovels(List<String> genreNames, Boolean isCompleted, Float novelRating,
-                                                       List<Integer> keywordIds, int page, int size) {
-        PageRequest pageRequest = PageRequest.of(page, size);
-        List<Genre> genres = getGenres(genreNames);
-        List<Keyword> keywords = getKeywords(keywordIds);
-
-        Page<Novel> novels = novelRepository.findFilteredNovels(pageRequest, genres, isCompleted, novelRating,
-                keywords);
-
-        List<NovelGetResponsePreview> novelGetResponsePreviews = novels.stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-
-        return FilteredNovelsGetResponse.of(novels.getTotalElements(), novels.hasNext(), novelGetResponsePreviews);
-    }
-
-    @Transactional(readOnly = true)
-    public SearchedNovelsGetResponse searchNovels(String query, int page, int size) {
-        PageRequest pageRequest = PageRequest.of(page, size);
-        String searchQuery = query.replaceAll("\\s+", "").replaceAll("[^a-zA-Z0-9가-힣]", "");
-
-        if (searchQuery.isBlank()) {
-            return SearchedNovelsGetResponse.of(0L, false, Collections.emptyList());
-        }
-
-        Page<Novel> novels = novelRepository.findSearchedNovels(pageRequest, searchQuery);
-
-        List<NovelGetResponsePreview> novelGetResponsePreviews = novels.stream()
-                .map(this::convertToDTO)
-                .collect(Collectors.toList());
-
-        return SearchedNovelsGetResponse.of(novels.getTotalElements(), novels.hasNext(), novelGetResponsePreviews);
-    }
-
-    private List<Genre> getGenres(List<String> genreNames) {
-        genreNames = genreNames == null
-                ? Collections.emptyList()
-                : genreNames;
-
-        List<Genre> genres = new ArrayList<>();
-        if (!genreNames.isEmpty()) {
-            for (String genreName : genreNames) {
-                genres.add(genreService.getGenreOrException(genreName));
-            }
-        }
-
-        return genres;
-    }
-
-    private List<Keyword> getKeywords(List<Integer> keywordIds) {
-        keywordIds = keywordIds == null
-                ? Collections.emptyList()
-                : keywordIds;
-
-        List<Keyword> keywords = new ArrayList<>();
-        if (!keywordIds.isEmpty()) {
-            for (Integer keywordId : keywordIds) {
-                keywords.add(keywordService.getKeywordOrException(keywordId));
-            }
-        }
-
-        return keywords;
-    }
-
-    private NovelGetResponsePreview convertToDTO(Novel novel) {
-        List<UserNovel> userNovels = novel.getUserNovels();
-
-        long interestCount = userNovels.stream()
-                .filter(UserNovel::getIsInterest)
-                .count();
-        long novelRatingCount = userNovels.stream()
-                .filter(un -> un.getUserNovelRating() != 0.0f)
-                .count();
-        double novelRatingSum = userNovels.stream()
-                .filter(un -> un.getUserNovelRating() != 0.0f)
-                .mapToDouble(UserNovel::getUserNovelRating)
-                .sum();
-
-        Float novelRatingAverage = novelRatingCount == 0
-                ? 0.0f
-                : Math.round((float) (novelRatingSum / novelRatingCount) * 10.0f) / 10.0f;
-
-        return NovelGetResponsePreview.of(
-                novel,
-                interestCount,
-                novelRatingAverage,
-                novelRatingCount
-        );
-    }
 
     @Transactional(readOnly = true)
     public PopularNovelsGetResponse getTodayPopularNovels() {
@@ -426,19 +110,4 @@ public class NovelService {
         return new PopularNovelsGetResponse(popularNovelResponses);
     }
 
-    @Transactional(readOnly = true)
-    public TasteNovelsGetResponse getTasteNovels(User user) {
-        List<Genre> preferGenres = genrePreferenceRepository.findByUser(user)
-                .stream()
-                .map(GenrePreference::getGenre)
-                .toList();
-
-        List<Novel> tasteNovels = userNovelRepository.findTasteNovels(preferGenres);
-
-        List<TasteNovelGetResponse> tasteNovelGetResponses = tasteNovels.stream()
-                .map(TasteNovelGetResponse::of)
-                .toList();
-
-        return TasteNovelsGetResponse.of(tasteNovelGetResponses);
-    }
 }
